+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New option:       gnssReceiverGeneratorStationNetwork: inputfileClock.
 - New option:       gnssReceiverGenerator: print preprocessing infos.
 - New option:       GroupPrograms: silently and additional outputfileLog.
+- Bugfix:           instrument files: empty files are now compatible to other instrument types.
 - Bugfix:           gnssProcessingStep: uninitialized normalEquationInfo.
 - Bugfix:           gnssProcessingStepForEachReceiverSeparately: variableReceiver was not set.
 - Bugfix:           gnssProcessingStepResolveAmbiguities: for writing empty ambiguity file.